pyproject.toml
[project]
name = "hippocampus"
version = "0.1.0"
description = "Shared biologically-inspired long-term memory for AI assistants (Devin, Claude Code, OpenCode, Windsurf, Antigravity)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "msk" }]
dependencies = [
    "mcp>=1.0.0",
    "python-ulid>=2.2.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
semantic = [
    "fastembed>=0.3.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
heavy = [
    "sentence-transformers>=3.0",
    "torch>=2.2",
    "einops>=0.7",
    "xformers>=0.0.27 ; sys_platform != 'darwin'",
]

[project.scripts]
hippo = "hippocampus.cli.main:cli"
hippocampus-mcp = "hippocampus.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hippocampus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
